requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
pillow>=8.3.1
opencv-python>=4.5.3
numpy>=1.21.2
scikit-learn>=0.24.2
transformers>=4.10.0
torch>=1.9.0